Keycodes aren't 16 bits anymore.
[git-p4: depot-paths = "//depot/qt/main/": change = 153201]

+++ src/gui/embedded/qkbdpc101_qws.h
@@ -21,7 +21,7 @@
 #ifndef QT_NO_QWS_KBD_PC101
 
 typedef struct QWSKeyMap {
-    ushort key_code;
+    uint key_code;
     ushort unicode;
     ushort shift_unicode;
     ushort ctrl_unicode;
